Rework system for getting song/artist data

+++ prebuild.sh
@@ -1,5 +1,6 @@
 cd server
-rm -rf ./spectre7
-git clone https://github.com/spectreseven1138/python-spectre7
-mv ./python-spectre7/spectre7 ./
-rm -rf ./python-spectre7
+if [ ! -d "./spectre7" ]; then
+  git clone https://github.com/spectreseven1138/python-spectre7
+  mv ./python-spectre7/spectre7 ./
+  rm -rf ./python-spectre7
+fi
